requirements.txt
appnope==0.1.0
APScheduler==3.6.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.3
oauth2client==4.1.3
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pygsheet==0.1.19b25
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
Telethon==1.17.5
terminado==0.9.1
testpath==0.4.4
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
